# api/src/main/resources/openapi.yml
openapi: 3.0.3
info:
  title: Desafío Técnico Gestión de Tareas
  description: API REST para la gestión de tareas con autenticación JWT.
  version: 1.0.0
servers:
  - url: http://localhost:8081
    description: Servidor Local
paths:
  /api/auth/login:
    post:
      tags:
        - Authentication
      summary: Autenticar usuario
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Credenciales inválidas

  /api/tasks:
    get:
      tags:
        - Task
      summary: Obtener todas las tareas (Filtrado opcional por estado y Paginado)
      operationId: getAllTasks
      parameters:
        - name: status
          in: query
          description: Nombre del estado para filtrar (PENDIENTE, EN_PROGRESO, etc.)
          required: false
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/TaskDTO'
                  totalElements:
                    type: integer
                  totalPages:
                    type: integer
    post:
      tags:
        - Task
      summary: Crear una nueva tarea
      operationId: createTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskRequest'
      responses:
        '201':
          description: Creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDTO'

  /api/tasks/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
    put:
      tags:
        - Task
      summary: Actualizar una tarea existente
      operationId: updateTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDTO'
    delete:
      tags:
        - Task
      summary: Eliminar una tarea
      operationId: deleteTask
      responses:
        '204':
          description: Sin contenido

components:
  schemas:
    AuthRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        type:
          type: string
          default: Bearer
    
    TaskRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        statusId:
          type: integer
          format: int64
      required:
        - title
        - statusId
    
    TaskDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        statusName:
          type: string
        username:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
  
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

security:
  - bearerAuth: []